requirements.txt
Flask==3.0.0
google-api-python-client==2.149.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
python-docx==1.1.0
gunicorn==21.2.0
